+++ IMS/Masters/Product.aspx.cs
@@ -10,11 +10,15 @@ using IMSBLL.EntityModel;
 using IMSBLL.DAL;
 using System.Configuration;
 using System.Reflection;
+using IMS.UserControl;
+
+
 
 namespace IMS
 {
-    public partial class clsProduct : System.Web.UI.Page
+    public partial class Product : System.Web.UI.Page
     {
+        
         /// <summary>
         /// All The objects That are used in coding
         /// </summary>
@@ -28,7 +32,6 @@ namespace IMS
         #endregion
         protected void Page_Load(object sender, EventArgs e)
         {
-
 
             try
             {
@@ -56,12 +59,12 @@ namespace IMS
         private void PageLoad()
         {
            
-            ddlcategorybind();
-            ddlunitbind();
-            ddlgodownbind();
-            ddltaxbind();
+            //ddlcategorybind();
+            //ddlunitbind();
+            //ddlgodownbind();
+            //ddltaxbind();
             loadDataTable();
-            ddlCategory.Focus();
+          //  ddlCategory.Focus();
         }
         private void SessionValue()
         {
@@ -69,7 +72,7 @@ namespace IMS
             companyId = Convert.ToInt32(Session["company_id"]);
             branchId = Convert.ToInt32(Session["branch_id"]);
         }
-        private void loadDataTable()
+        public void loadDataTable()
         {
 
             try
@@ -171,257 +174,257 @@ namespace IMS
         }
 
 
-        public void ddlcategorybind()
-        {
+        //public void ddlcategorybind()
+        //{
             
-            List<tbl_category> cd = context.tbl_category.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).ToList();
-            ddlCategory.DataTextField = "category_name";
-            ddlCategory.DataValueField = "category_id";
-            ddlCategory.DataSource = cd;
-            ddlCategory.DataBind();
-            ddlCategory.Items.Insert(0, new ListItem("--Select Category--", "0"));
-        }
-        public void ddlunitbind()
-        {
-            List<tbl_unit> cd = context.tbl_unit.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).ToList();
-            ddlUnit.DataTextField = "unit_name";
-            ddlUnit.DataValueField = "unit_id";
-            ddlUnit.DataSource = cd;
-            ddlUnit.DataBind();
-            ddlUnit.Items.Insert(0, new ListItem("--Select Unit--", "0"));
-        }
+        //    List<tbl_category> cd = context.tbl_category.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).ToList();
+        //    ddlCategory.DataTextField = "category_name";
+        //    ddlCategory.DataValueField = "category_id";
+        //    ddlCategory.DataSource = cd;
+        //    ddlCategory.DataBind();
+        //    ddlCategory.Items.Insert(0, new ListItem("--Select Category--", "0"));
+        //}
+        //public void ddlunitbind()
+        //{
+        //    List<tbl_unit> cd = context.tbl_unit.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).ToList();
+        //    ddlUnit.DataTextField = "unit_name";
+        //    ddlUnit.DataValueField = "unit_id";
+        //    ddlUnit.DataSource = cd;
+        //    ddlUnit.DataBind();
+        //    ddlUnit.Items.Insert(0, new ListItem("--Select Unit--", "0"));
+        //}
 
-        public void ddlgodownbind()
-        {
-            List<tbl_godown> gd = context.tbl_godown.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).ToList();
+        //public void ddlgodownbind()
+        //{
+        //    List<tbl_godown> gd = context.tbl_godown.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).ToList();
 
-            ddlGodown.DataTextField = "godown_name";
-            ddlGodown.DataValueField = "godown_id";
-            ddlGodown.DataSource = gd;
-            ddlGodown.DataBind();
-            ddlGodown.Items.Insert(0, new ListItem("--Select Godown--", "0"));
-        }
-        public void ddlrackbind(int godownid)
-        {
-            int g_id = Convert.ToInt32(ddlGodown.SelectedValue);
-            List<tbl_rack> cd = context.tbl_rack.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId && x.godown_id==godownid).ToList();
-            ddlRack.DataTextField = "rack_name";
-            ddlRack.DataValueField = "rack_id";
-            ddlRack.DataSource = cd;
-            ddlRack.DataBind();
-            ddlRack.Items.Insert(0, new ListItem("--Select Rack--", "0"));
-            if (cd.Count == 0)
-            {
-                lblrackerror.Text = "No rack found ";
-            }
-            else
-            {
-                lblrackerror.Text = string.Empty;
-            }
-        }
-        public void ddltaxbind()
-        {
-            List<tbl_taxgroup> cd = context.tbl_taxgroup.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).ToList();
-            dataTable = ToDataTable(cd);
-            ddlTaxgroup.DataTextField = "group_name";
-            ddlTaxgroup.DataValueField = "group_id";
-            ddlTaxgroup.DataSource = cd;
-            ddlTaxgroup.DataBind();
-        }
+        //    ddlGodown.DataTextField = "godown_name";
+        //    ddlGodown.DataValueField = "godown_id";
+        //    ddlGodown.DataSource = gd;
+        //    ddlGodown.DataBind();
+        //    ddlGodown.Items.Insert(0, new ListItem("--Select Godown--", "0"));
+        //}
+        //public void ddlrackbind(int godownid)
+        //{
+        //    int g_id = Convert.ToInt32(ddlGodown.SelectedValue);
+        //    List<tbl_rack> cd = context.tbl_rack.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId && x.godown_id==godownid).ToList();
+        //    ddlRack.DataTextField = "rack_name";
+        //    ddlRack.DataValueField = "rack_id";
+        //    ddlRack.DataSource = cd;
+        //    ddlRack.DataBind();
+        //    ddlRack.Items.Insert(0, new ListItem("--Select Rack--", "0"));
+        //    if (cd.Count == 0)
+        //    {
+        //        lblrackerror.Text = "No rack found ";
+        //    }
+        //    else
+        //    {
+        //        lblrackerror.Text = string.Empty;
+        //    }
+        //}
+        //public void ddltaxbind()
+        //{
+        //    List<tbl_taxgroup> cd = context.tbl_taxgroup.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).ToList();
+        //    dataTable = ToDataTable(cd);
+        //    ddlTaxgroup.DataTextField = "group_name";
+        //    ddlTaxgroup.DataValueField = "group_id";
+        //    ddlTaxgroup.DataSource = cd;
+        //    ddlTaxgroup.DataBind();
+        //}
 
-        private void savedcode()
-        {
-            try
-            {
-                if (hd.Value == "false")
-                {
-                    if (hde.Value == "false")
-                    {
-                        Label1.Text = String.Empty;
-                        lblcheckDoubleError.Text = String.Empty;
-                        tbl_product product = new tbl_product();
-                        product.company_id = companyId;
-                        product.branch_id = branchId;
-                        product.category_id = Int32.Parse(ddlCategory.SelectedValue);
-                        product.unit_id = Int32.Parse(ddlUnit.SelectedValue);
-                        product.godown_id = Int32.Parse(ddlGodown.SelectedValue);
-                        product.rack_id = Int32.Parse(ddlRack.SelectedValue);
-                        //product.tax_id = Int32.Parse(ddlTax.SelectedValue);
-                        product.product_name = txtProductName.Text;
-                        product.product_code = txtProductCode.Text;
-                        product.hsn_code = txtHSNCode.Text;
-                        product.reorder_level = Int32.Parse(txtReorderqty.Text);
-                        product.purchas_price = decimal.Parse(txtPurchasePrice.Text);
-                        product.sales_price = decimal.Parse(txtSalesPrice.Text);
-                        product.created_by = User_id;
-                        product.created_date = DateTime.Today;
-                        product.status = true;
-                        product.modified_by = "";
-                        product.modified_date = null;
-                        ////Shakeeb
-                        ////pd.Insert(pd);
-                        //Entity Framework Saving Awais
+        //private void savedcode()
+        //{
+        //    try
+        //    {
+        //        if (hd.Value == "false")
+        //        {
+        //            if (hde.Value == "false")
+        //            {
+        //                Label1.Text = String.Empty;
+        //                lblcheckDoubleError.Text = String.Empty;
+        //                tbl_product product = new tbl_product();
+        //                product.company_id = companyId;
+        //                product.branch_id = branchId;
+        //                product.category_id = Int32.Parse(ddlCategory.SelectedValue);
+        //                product.unit_id = Int32.Parse(ddlUnit.SelectedValue);
+        //                product.godown_id = Int32.Parse(ddlGodown.SelectedValue);
+        //                product.rack_id = Int32.Parse(ddlRack.SelectedValue);
+        //                //product.tax_id = Int32.Parse(ddlTax.SelectedValue);
+        //                product.product_name = txtProductName.Text;
+        //                product.product_code = txtProductCode.Text;
+        //                product.hsn_code = txtHSNCode.Text;
+        //                product.reorder_level = Int32.Parse(txtReorderqty.Text);
+        //                product.purchas_price = decimal.Parse(txtPurchasePrice.Text);
+        //                product.sales_price = decimal.Parse(txtSalesPrice.Text);
+        //                product.created_by = User_id;
+        //                product.created_date = DateTime.Today;
+        //                product.status = true;
+        //                product.modified_by = "";
+        //                product.modified_date = null;
+        //                ////Shakeeb
+        //                ////pd.Insert(pd);
+        //                //Entity Framework Saving Awais
 
-                        foreach (ListItem item in ddlTaxgroup.Items)
-                        {
+        //                foreach (ListItem item in ddlTaxgroup.Items)
+        //                {
 
-                            if (item.Selected)
-                            {
-                                tbl_productTaxGroup productTax = new tbl_productTaxGroup();
-                                productTax.group_id = Convert.ToInt32(item.Value);
-                                productTax.isSelected = true;
-                                product.tbl_productTaxGroup.Add(productTax);
-                            }
+        //                    if (item.Selected)
+        //                    {
+        //                        tbl_productTaxGroup productTax = new tbl_productTaxGroup();
+        //                        productTax.group_id = Convert.ToInt32(item.Value);
+        //                        productTax.isSelected = true;
+        //                        product.tbl_productTaxGroup.Add(productTax);
+        //                    }
 
-                        }
+        //                }
 
-                        context.tbl_product.Add(product);
-                        context.SaveChanges();
-                        loadDataTable();
-                        clr();
-                        divalert.Visible = true;
-                        lblAlert.Text = "Product Saved Successfully ";
+        //                context.tbl_product.Add(product);
+        //                context.SaveChanges();
+        //                loadDataTable();
+        //                clr();
+        //                divalert.Visible = true;
+        //                lblAlert.Text = "Product Saved Successfully ";
 
 
                         
-                        //context.tbl_product.Add(pd);
-                        //context.SaveChanges();
-                    }
-                    else
-                    {
-                        divalert.Visible = false;
-                        lblcheckDoubleError.ForeColor = System.Drawing.Color.Red;
-                        lblcheckDoubleError.Text = "This Product Code name already Exists";
-                    }
+        //                //context.tbl_product.Add(pd);
+        //                //context.SaveChanges();
+        //            }
+        //            else
+        //            {
+        //                divalert.Visible = false;
+        //                lblcheckDoubleError.ForeColor = System.Drawing.Color.Red;
+        //                lblcheckDoubleError.Text = "This Product Code name already Exists";
+        //            }
 
-                }
-                else
-                {
-                    divalert.Visible = false;
-                    Label1.ForeColor = System.Drawing.Color.Red;
-                    Label1.Text = "This Product name already Exists";
-                }
-            }
-            catch (Exception ex)
-            {
+        //        }
+        //        else
+        //        {
+        //            divalert.Visible = false;
+        //            Label1.ForeColor = System.Drawing.Color.Red;
+        //            Label1.Text = "This Product name already Exists";
+        //        }
+        //    }
+        //    catch (Exception ex)
+        //    {
 
-                ErrorLog.saveerror(ex);
-                //Do Logging
-            }
-        }
+        //        ErrorLog.saveerror(ex);
+        //        //Do Logging
+        //    }
+        //}
 
-        private void updatecode()
-        {
-            try
-            {
-                if (hd.Value != "true")
-                {
-                    if (hde.Value != "true")
-                    {
-                        Label1.Text = string.Empty;
-                        lblcheckDoubleError.Text = String.Empty;
-                        GridViewRow row = GridView1.SelectedRow;
-                        int product_id = Convert.ToInt32(ViewState["productId"]);
-                        decimal sales_price = decimal.Parse(txtSalesPrice.Text);
-                        decimal purchase_price = decimal.Parse(txtPurchasePrice.Text);
-                        int cat_id = Int32.Parse(ddlCategory.SelectedValue);
-                        int unit_id = Int32.Parse(ddlUnit.SelectedValue);
-                        int godown_id = Int32.Parse(ddlGodown.SelectedValue);
-                        int rack_id = Int32.Parse(ddlRack.SelectedValue);
-                        //int tax_id = Int32.Parse(ddlTax.SelectedValue);
+        //private void updatecode()
+        //{
+        //    try
+        //    {
+        //        if (hd.Value != "true")
+        //        {
+        //            if (hde.Value != "true")
+        //            {
+        //                Label1.Text = string.Empty;
+        //                lblcheckDoubleError.Text = String.Empty;
+        //                GridViewRow row = GridView1.SelectedRow;
+        //                int product_id = Convert.ToInt32(ViewState["productId"]);
+        //                decimal sales_price = decimal.Parse(txtSalesPrice.Text);
+        //                decimal purchase_price = decimal.Parse(txtPurchasePrice.Text);
+        //                int cat_id = Int32.Parse(ddlCategory.SelectedValue);
+        //                int unit_id = Int32.Parse(ddlUnit.SelectedValue);
+        //                int godown_id = Int32.Parse(ddlGodown.SelectedValue);
+        //                int rack_id = Int32.Parse(ddlRack.SelectedValue);
+        //                //int tax_id = Int32.Parse(ddlTax.SelectedValue);
                         
-                        //context.sp_UpdateProduct(companyId, branchId, product_id, cat_id, rack_id, godown_id, tax_id, unit_id, orderlevel, purchase_price, sales_price, txtProductCode.Text, txtHSNCode.Text, txtProductName.Text, User_id, DateTime.Today);
+        //                //context.sp_UpdateProduct(companyId, branchId, product_id, cat_id, rack_id, godown_id, tax_id, unit_id, orderlevel, purchase_price, sales_price, txtProductCode.Text, txtHSNCode.Text, txtProductName.Text, User_id, DateTime.Today);
 
-                        var product = context.tbl_product.Where(w => w.product_id == product_id).FirstOrDefault();
-                        product.product_name = txtProductName.Text;
-                        product.product_code = txtProductCode.Text;
-                        product.hsn_code = txtHSNCode.Text;
-                        product.purchas_price = decimal.Parse(txtPurchasePrice.Text);
-                        product.reorder_level = Convert.ToInt32(txtReorderqty.Text);
-                        product.sales_price = decimal.Parse(txtSalesPrice.Text);
-                        product.rack_id = rack_id;
-                        product.category_id = cat_id;
-                        product.godown_id = godown_id;
+        //                var product = context.tbl_product.Where(w => w.product_id == product_id).FirstOrDefault();
+        //                product.product_name = txtProductName.Text;
+        //                product.product_code = txtProductCode.Text;
+        //                product.hsn_code = txtHSNCode.Text;
+        //                product.purchas_price = decimal.Parse(txtPurchasePrice.Text);
+        //                product.reorder_level = Convert.ToInt32(txtReorderqty.Text);
+        //                product.sales_price = decimal.Parse(txtSalesPrice.Text);
+        //                product.rack_id = rack_id;
+        //                product.category_id = cat_id;
+        //                product.godown_id = godown_id;
 
-                        //Delete Existing Mapping and add new mapping for product and tax group
-                        //var tbl_productTaxGroup = new tbl_productTaxGroup { product_id = product_id};
-                        //context.tbl_productTaxGroup.Attach(tbl_productTaxGroup);
-                        //context.tbl_productTaxGroup.Remove(tbl_productTaxGroup);
-                        //context.SaveChanges();
-                        foreach (var grpType in context.tbl_productTaxGroup.Where(w => w.product_id == product_id).ToList())
-                        {
-                            product.tbl_productTaxGroup.Remove(grpType);
-                        }
+        //                //Delete Existing Mapping and add new mapping for product and tax group
+        //                //var tbl_productTaxGroup = new tbl_productTaxGroup { product_id = product_id};
+        //                //context.tbl_productTaxGroup.Attach(tbl_productTaxGroup);
+        //                //context.tbl_productTaxGroup.Remove(tbl_productTaxGroup);
+        //                //context.SaveChanges();
+        //                foreach (var grpType in context.tbl_productTaxGroup.Where(w => w.product_id == product_id).ToList())
+        //                {
+        //                    product.tbl_productTaxGroup.Remove(grpType);
+        //                }
 
-                        for (int i = 0; i < ddlTaxgroup.Items.Count; i++)
-                        {
+        //                for (int i = 0; i < ddlTaxgroup.Items.Count; i++)
+        //                {
                                                     
                             
-                            if (ddlTaxgroup.Items[i].Selected)
-                            {
+        //                    if (ddlTaxgroup.Items[i].Selected)
+        //                    {
                                 
-                                tbl_productTaxGroup productTaxGroup = new tbl_productTaxGroup();
-                                ListItem item = ddlTaxgroup.Items[i];
-                                int groupId = int.Parse(ddlTaxgroup.Items[i].Value);
-                                productTaxGroup.product_id = product_id;
-                                productTaxGroup.group_id = groupId;
-                                product.tbl_productTaxGroup.Add(productTaxGroup);
-                            }
-                        }
+        //                        tbl_productTaxGroup productTaxGroup = new tbl_productTaxGroup();
+        //                        ListItem item = ddlTaxgroup.Items[i];
+        //                        int groupId = int.Parse(ddlTaxgroup.Items[i].Value);
+        //                        productTaxGroup.product_id = product_id;
+        //                        productTaxGroup.group_id = groupId;
+        //                        product.tbl_productTaxGroup.Add(productTaxGroup);
+        //                    }
+        //                }
                        
 
-                        product.unit_id = unit_id;
-                        product.modified_by = User_id;
-                        product.modified_date = DateTime.Now;
+        //                product.unit_id = unit_id;
+        //                product.modified_by = User_id;
+        //                product.modified_date = DateTime.Now;
 
 
-                        context.SaveChanges();
+        //                context.SaveChanges();
 
 
 
-                        btnUpdate.Visible = false;
-                        btnSave.Visible = true;
-                        divalert.Visible = true;
-                        clr();
-                        loadDataTable();
-                        lblAlert.Text = "Product Updated Successfully ";
-                    }
-                    else
-                    {
-                        divalert.Visible = false;
-                        lblcheckDoubleError.ForeColor = System.Drawing.Color.Red;
-                        lblcheckDoubleError.Text = "This Product Code name already Exists";
-                    }
-                }
-                else
-                {
-                    divalert.Visible = false;
-                    Label1.ForeColor = System.Drawing.Color.Red;
-                    Label1.Text = "This Product Name already Exists";
-                }
-            }
-            catch (Exception ex)
-            {
+        //                btnUpdate.Visible = false;
+        //                btnSave.Visible = true;
+        //                divalert.Visible = true;
+        //                clr();
+        //                loadDataTable();
+        //                lblAlert.Text = "Product Updated Successfully ";
+        //            }
+        //            else
+        //            {
+        //                divalert.Visible = false;
+        //                lblcheckDoubleError.ForeColor = System.Drawing.Color.Red;
+        //                lblcheckDoubleError.Text = "This Product Code name already Exists";
+        //            }
+        //        }
+        //        else
+        //        {
+        //            divalert.Visible = false;
+        //            Label1.ForeColor = System.Drawing.Color.Red;
+        //            Label1.Text = "This Product Name already Exists";
+        //        }
+        //    }
+        //    catch (Exception ex)
+        //    {
 
-                ErrorLog.saveerror(ex);
-                //Do Logging
-            }
-        }
+        //        ErrorLog.saveerror(ex);
+        //        //Do Logging
+        //    }
+        //}
 
-        public void clr()
-        {
-            txtHSNCode.Text = string.Empty;
-            txtProductCode.Text = string.Empty;
-            txtProductName.Text = string.Empty;
-            txtPurchasePrice.Text = string.Empty;
-            txtReorderqty.Text = string.Empty;
-            txtSalesPrice.Text = string.Empty;
-            ddlCategory.SelectedIndex = 0;
-            ddlGodown.SelectedIndex = 0;
-            ddlRack.SelectedIndex = 0;
-            ddlTaxgroup.SelectedIndex = 0;
-            ddlUnit.SelectedIndex = 0;
-        }
+        //public void clr()
+        //{
+        //    txtHSNCode.Text = string.Empty;
+        //    txtProductCode.Text = string.Empty;
+        //    txtProductName.Text = string.Empty;
+        //    txtPurchasePrice.Text = string.Empty;
+        //    txtReorderqty.Text = string.Empty;
+        //    txtSalesPrice.Text = string.Empty;
+        //    ddlCategory.SelectedIndex = 0;
+        //    ddlGodown.SelectedIndex = 0;
+        //    ddlRack.SelectedIndex = 0;
+        //    ddlTaxgroup.SelectedIndex = 0;
+        //    ddlUnit.SelectedIndex = 0;
+        //}
 
         #endregion
         /// <summary>
@@ -429,30 +432,42 @@ namespace IMS
         /// </summary>
 
         #region Events
-        protected void btnSave_Click(object sender, EventArgs e)
-        {
-            savedcode();
-        }
+        //protected void btnSave_Click(object sender, EventArgs e)
+        //{
+        //    //savedcode();
+        //}
 
         
         protected void GridView1_SelectedIndexChanged(object sender, EventArgs e)
         {
             try
             {
-                Label1.Text = String.Empty;
-                lblcheckDoubleError.Text = String.Empty;
+                //Label1.Text = String.Empty;
+                //lblcheckDoubleError.Text = String.Empty;
                 GridViewRow row = GridView1.SelectedRow;
                 int productId = Convert.ToInt32(GridView1.DataKeys[row.RowIndex].Value);
-                ViewState["productId"] = productId;
+
+                Session["productId"] = productId;
+                ctrlProduct.productid = productId.ToString();
+                
+
+
+                ctrlProduct.purchprice = Convert.ToDecimal(row.Cells[3].Text);
+                ctrlProduct.saleprice = Convert.ToDecimal(row.Cells[2].Text);
+                ctrlProduct.productname = row.Cells[0].Text;
+                ctrlProduct.productcode = row.Cells[4].Text;
+                ctrlProduct.HSNcode = row.Cells[10].Text;
+                ctrlProduct.reorderlevel = Convert.ToInt32(row.Cells[5].Text);
                 //int taxId = GridView1.SelectedIndex;
-                ddlCategory.SelectedValue=row.Cells[6].Text;
-                ddlUnit.SelectedValue = (row.Cells[9].Text);
-                ddlGodown.SelectedValue = (row.Cells[7].Text);
-                ddlRack.SelectedValue = (row.Cells[8].Text);
+                ctrlProduct.categoryid = Convert.ToInt32(row.Cells[6].Text);
+                ctrlProduct.unitid = Convert.ToInt32(row.Cells[9].Text);
+                ctrlProduct.godownid = Convert.ToInt32(row.Cells[7].Text);
+                ctrlProduct.rackid = Convert.ToInt32(row.Cells[8].Text);
                // ddlTax.SelectedValue = (row.Cells[13].Text);
 
+                List<tbl_taxgroup> cd = context.tbl_taxgroup.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).ToList();
+                dataTable = ToDataTable(cd);
 
-                ddltaxbind();
                 if (productId != 0)
                 {
                     List<tbl_productTaxGroup> selectedItem = context.tbl_productTaxGroup.Where(p => p.product_id == productId).ToList();
@@ -469,9 +484,9 @@ namespace IMS
                                 int group_id = Convert.ToInt32(dataTable.Rows[i]["group_id"]);
                                 if (group_id == groupID)
                                 {
-                                    //ddlTaxgroup.SelectedValue = groupID.ToString();
-                                    //ddlTaxgroup.Items[j].Value = groupID.ToString();
-                                    ddlTaxgroup.Items[i].Selected = true;
+                                    ctrlProduct.Taxdropdown.SelectedValue = groupID.ToString();
+                                    ctrlProduct.Taxdropdown.Items[j].Value = groupID.ToString();
+                                    ctrlProduct.Taxdropdown.Items[i].Selected = true;
                                     break;
                                 }
 
@@ -482,18 +497,13 @@ namespace IMS
                 }
 
 
-                txtPurchasePrice.Text = row.Cells[3].Text;
-                txtSalesPrice.Text = row.Cells[2].Text;
-                txtProductName.Text = row.Cells[0].Text;
-                txtProductCode.Text = row.Cells[4].Text;
-                txtHSNCode.Text = row.Cells[10].Text;
-                txtReorderqty.Text = row.Cells[5].Text;
-                btnSave.Visible = false;
-                btnUpdate.Visible = true;
+               
+                //btnSave.Visible = false;
+                //btnUpdate.Visible = true;
                 divalert.Visible = false;
-                ddlCategory.Focus();
-                hde.Value = string.Empty;
-                hd.Value = string.Empty;
+                //ddlCategory.Focus();
+                //hde.Value = string.Empty;
+                //hd.Value = string.Empty;
             }
             catch (Exception ex)
             {
@@ -504,10 +514,10 @@ namespace IMS
 
 
         }
-        protected void btnUpdate_Click(object sender, EventArgs e)
-        {
-            updatecode();
-        }
+        //protected void btnUpdate_Click(object sender, EventArgs e)
+        //{
+        //    updatecode();
+        //}
 
         protected void GridView1_RowCommand(object sender, GridViewCommandEventArgs e)
         {
@@ -539,7 +549,7 @@ namespace IMS
                 context.sp_DeleteProduct(companyId, rowIndex);
                 divalert.Visible = true;
                 lblAlert.Text = "Product Deleted Successfully ";
-                clr();
+               // clr();
                 loadDataTable();
             }
             catch (Exception ex)
@@ -554,14 +564,14 @@ namespace IMS
         {
             try
             {
-                lblcheckDoubleError.Text = String.Empty;
-                btnSave.Visible = true;
-                btnUpdate.Visible = false;
+                //lblcheckDoubleError.Text = String.Empty;
+                //btnSave.Visible = true;
+                //btnUpdate.Visible = false;
                 divalert.Visible = false;
-                clr();
+              //  clr();
                 ViewState["gridrow"] = null;
-                hde.Value = string.Empty;
-                hd.Value = string.Empty;
+                //hde.Value = string.Empty;
+                //hd.Value = string.Empty;
             }
             catch (Exception ex)
             {
@@ -569,28 +579,28 @@ namespace IMS
             }
         }
 
-        protected void Button1_Click(object sender, EventArgs e)
-        {
-            clr();
-        }
+        //protected void Button1_Click(object sender, EventArgs e)
+        //{
+        //    clr();
+        //}
 
-        protected void ddlGodown_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            try
-            {
-                //lblrackerror.Text = string.Empty;
-                int g_id = Convert.ToInt32(ddlGodown.SelectedValue);
-                ddlrackbind(g_id);
-                UpdatePanel1.Update();
+        //protected void ddlGodown_SelectedIndexChanged(object sender, EventArgs e)
+        //{
+        //    try
+        //    {
+        //        //lblrackerror.Text = string.Empty;
+        //        int g_id = Convert.ToInt32(ddlGodown.SelectedValue);
+        //        ddlrackbind(g_id);
+        //        UpdatePanel1.Update();
 
-            }
-            catch (Exception ex)
-            {
+        //    }
+        //    catch (Exception ex)
+        //    {
 
-                ErrorLog.saveerror(ex);
-                //Do Logging
-            }
-        }
+        //        ErrorLog.saveerror(ex);
+        //        //Do Logging
+        //    }
+        //}
 
         protected void GridView1_PreRender(object sender, EventArgs e)
         {
@@ -627,43 +637,23 @@ namespace IMS
         }
         #endregion
 
-        protected void btnCloseMode_Click(object sender, EventArgs e)
-        {
-            try
-            {
-                ddlcategorybind();
-                ddlgodownbind();
-                ddlrackbind(Convert.ToInt32(ddlGodown.SelectedValue));
-                ddltaxbind();
-                ddlunitbind();
-                //ddlVendorbind();
-                //ddlbatchbind();
-                ScriptManager.RegisterStartupScript(this, GetType(), "Close Modal Popup", "Closepopup();", true);
-                UpdatePanel1.Update();
-            }
-            catch (Exception ex)
-            {
+     
 
-                ErrorLog.saveerror(ex);
-                //Do Logging
-            }
-        }
+        //protected void btncloseofmodelcategory_ServerClick(object sender, EventArgs e)
+        //{
+        //    // on close of popup below code will select last inserted category in category dropdown
+        //    var set = context.tbl_category.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).OrderByDescending(x => x.category_id).FirstOrDefault();
+        //    ddlcategorybind();
+        //    ddlCategory.SelectedValue = set.category_id.ToString();
+        //}
 
-        protected void btncloseofmodelcategory_ServerClick(object sender, EventArgs e)
-        {
-            // on close of popup below code will select last inserted category in category dropdown
-            var set = context.tbl_category.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).OrderByDescending(x => x.category_id).FirstOrDefault();
-            ddlcategorybind();
-            ddlCategory.SelectedValue = set.category_id.ToString();
-        }
-
-        protected void btnunitmodelclose_ServerClick(object sender, EventArgs e)
-        {
-            // on close of popup below code will select last inserted Unit in unit dropdown
-            var set = context.tbl_unit.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).OrderByDescending(x => x.unit_id).FirstOrDefault();
-            ddlunitbind();
-            ddlUnit.SelectedValue = set.unit_id.ToString();
-        }
+        //protected void btnunitmodelclose_ServerClick(object sender, EventArgs e)
+        //{
+        //    // on close of popup below code will select last inserted Unit in unit dropdown
+        //    var set = context.tbl_unit.Where(x => x.status == true && x.company_id == companyId && x.branch_id == branchId).OrderByDescending(x => x.unit_id).FirstOrDefault();
+        //    ddlunitbind();
+        //    ddlUnit.SelectedValue = set.unit_id.ToString();
+        //}
